requirements.txt
# requirements.txt

PyQt5==5.15.11
matplotlib==3.10.3
numpy==2.1.3
pennylane==0.41.1
pennylane-lightning==0.41.1
scikit-learn==1.6.1
tensorflow==2.19.0
ta==0.11.0
pandas==2.2.3
yfinance==0.2.59
python-dotenv==1.1.0
requests==2.32.3
alpha-vantage==3.0.0

# New dependencies for improvements
tenacity>=8.2.0
PyYAML>=6.0
